AIDetector/code/backend/backend-code/requirements-dev.txt
django==4.1.1
django-cors-headers
djangorestframework
djangorestframework-simplejwt
celery
pymysql
psycopg2-binary
pillow
numpy
requests
reportlab
sshtunnel
